Add new Razor-Tag with option z
We use sed to modify the tizu file, but sed has no "non greedy" search/replace.
So as a kind of workaround, we have to have unique end tags, which results in
start and end tags are looking the same.

The razer option need a couple of special handling. Where all other options
only ever select one of the supported possibilities, several razor options
can be active at the same time.

This seems to work, but I'm not sure if the implementation is well done.