Fix/Add missing \e issue escape code
Unlike the \S code, the \e code may have been deliberately omitted eight
years ago, because there is no reason to use it. But now, not just for
the sake of completeness, but because there was a change in the agetty
code where color names are supported and not just “raw codes”. These are
really nice and human friendly.

  karelzak committed on Feb 17, 2016
  agetty: support \e{name} for issue file
  https://github.com/util-linux/util-linux/commit/d689166b11c5ee89a754d46cc1ce09f27bde0a31

Hm, interesting. Only a few weeks after first release of tulizu there is
this change that offers a similar color handling.